fix: permissive linebreaking + refactor processings
this time, the spoiler, the blockquoting and the permissive linebreaking
are well-embedded (unless with the old library)
